% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fixtures.R
\name{get_fixtures}
\alias{get_fixtures}
\title{Fixtures from top 5 european leagues}
\usage{
get_fixtures(league_name)
}
\arguments{
\item{league_name}{must be a character league code name}
}
\value{
Data frame of fixtures from a defined league
}
\description{
Fixtures from top 5 european leagues
}
\details{
League code name must be one of "EPL", "BundesLiga", "LaLiga", "Ligue1", "SerieA"
}
\examples{
get_fixtures("EPL")


}
