% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Results from top 5 european leagues}
\usage{
get_results(league_name)
}
\arguments{
\item{league_name}{must be a character league code name}
}
\value{
Data frame of results from a defined league
}
\description{
Results from top 5 european leagues
}
\details{
League code name must be one of "EPL", "BundesLiga", "LaLiga", "Ligue1", "SerieA"
}
\examples{
get_results("EPL")


}
