% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scorers.R
\name{get_scorers}
\alias{get_scorers}
\title{Goal scorers frop top 5 european leagues}
\usage{
get_scorers(league_name)
}
\arguments{
\item{league_name}{must be a character league code name}
}
\value{
Data frame of goal scorers from a defined league
}
\description{
Goal scorers frop top 5 european leagues
}
\details{
League code name must be one of "EPL", "BundesLiga", "LaLiga", "Ligue1", "SerieA"
}
\examples{
get_scorers("EPL")


}
