% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standings.R
\name{get_standings}
\alias{get_standings}
\title{Standings from top 5 european leagues}
\usage{
get_standings(league_name = "EPL")
}
\arguments{
\item{league_name}{must be a character league code name}
}
\value{
Data frame of results from a defined league
}
\description{
Standings from top 5 european leagues
}
\details{
League code name must be one of "EPL", "BundesLiga", "LaLiga", "Ligue1", "SerieA"
}
\examples{
get_standings("EPL")


}
